LPD-18678 Move PublishScheduler and ScheduledPublishInfo out of web module and reconfigure implementation
