+++ modules/apps/change-tracking/change-tracking-api/src/main/java/com/liferay/change/tracking/scheduler/ScheduledPublishInfo.java
@@ -1,9 +1,9 @@
 /**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.change.tracking.web.internal.scheduler;
+package com.liferay.change.tracking.scheduler;
 
 import com.liferay.change.tracking.model.CTCollection;
 
@@ -13,6 +13,16 @@ import java.util.Date;
  * @author Preston Crary
  */
 public class ScheduledPublishInfo {
+
+	public ScheduledPublishInfo(
+		CTCollection ctCollection, long userId, String jobName,
+		Date startDate) {
+
+		_ctCollection = ctCollection;
+		_userId = userId;
+		_jobName = jobName;
+		_startDate = startDate;
+	}
 
 	public CTCollection getCTCollection() {
 		return _ctCollection;
@@ -28,16 +38,6 @@ public class ScheduledPublishInfo {
 
 	public long getUserId() {
 		return _userId;
-	}
-
-	protected ScheduledPublishInfo(
-		CTCollection ctCollection, long userId, String jobName,
-		Date startDate) {
-
-		_ctCollection = ctCollection;
-		_userId = userId;
-		_jobName = jobName;
-		_startDate = startDate;
 	}
 
 	private final CTCollection _ctCollection;
