+++ modules/apps/change-tracking/change-tracking-api/src/main/resources/com/liferay/change/tracking/scheduler/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
