+++ modules/apps/change-tracking/change-tracking-web/src/main/java/com/liferay/change/tracking/web/internal/display/context/ReschedulePublicationDisplayContext.java
@@ -6,7 +6,7 @@
 package com.liferay.change.tracking.web.internal.display.context;
 
 import com.liferay.change.tracking.model.CTCollection;
-import com.liferay.change.tracking.web.internal.scheduler.ScheduledPublishInfo;
+import com.liferay.change.tracking.scheduler.ScheduledPublishInfo;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.language.Language;
